pyproject.toml
[project]
name = "atotw-mcp"
version = "0.1.0"
description = "MCP server for Anaesthesia Tutorial of the Week content"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pypdf>=4.0.0",
    "chromadb>=0.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=0.1.0",
]

[project.scripts]
sync = "sync:main"
mcp-server = "mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]